requirements.txt
# Telegram Bot
python-telegram-bot[job-queue]==22.5

# Database
sqlalchemy>=2.0.0
asyncpg>=0.30.0  # PostgreSQL async driver
aiosqlite>=0.20.0  # для локальной разработки

# Google Drive & Sheets
google-api-python-client>=2.150.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
gspread>=6.0.0

# Utils
python-dotenv>=1.0.0
loguru>=0.7.0
httpx>=0.27.0  # HTTP клиент для DaData API